Completely revamp image extraction
Before, nsbox was manually grabbing each layer and extracting it by hand.
However, this did not take whiteout files into consideration, which
became an issue once the Fedora image moved to removing the masked
service files in a layer: now, the masked files remained in the resulting
container.

To remedy this, go-containerregistry's built-in mutate.Extract is used
instead. This also moves from archiver to another untar library, partly
because archiver doesn't support a full extraction from an archive.

Closes #7, #9, #10.