Avoid passing extra arguments during a redirect
Before, VisitAll was used to iterate over all arguments and forward
them during a redirect. That way, -workdir would always be included,
so when pkexec changed directories, nsbox would change back to the
original working directory.

However, 'nsbox config' checks if -virtual-network and -share-cgroups
were passed and, if so, it updates their saved values with the CLI
ones. Since the redirector was iterating over *all* arguments, it
would explicily pass -virtual-network and -share-cgroups, meaning
that they would always be toggled on in 'nsbox config' even if not
given.

The fix is to simply use Visit instead and manually / explicitly
forward -workdir.

Fixes #C7.