Properly handle modern Silverblue /home locations
Classic Silverblue installs have $HOME=/home/USER where /home is a
symlink, so the nsbox code handled that case but generalized it to
any home parent that is a symlink.

However, modern Silverblue has $HOME=/var/home/USER, but /home is still
a symlink. Therefore, we need to make an equivalent symlink inside the
container.

In order to achieve this, a lot of the generalization was pulled out of
the code, as I'm not aware of another platform that does this and not
many distros are officially supported anyway. It also allowed for some
overall code simplification to take place.

Closes #C4.